饱和加法指令似乎还有些问题，改了。

+++ arm/src/inst/qadd16.c
@@ -22,10 +22,10 @@ int32 arm_inst_qadd16(cpu_state_t *st, uint32 inst)
 
 	//低位结果
 	int32 resL = (int32)(int16)bitm(regv(rm), 0, 15) + (int16)bitm(regv(rn), 0, 15);
-	uint32 L = signed_sat(resL, 16, &issat);
+	uint16 L = signed_sat(resL, 16, &issat);
 	//高位结果
 	int32 resH = (int32)(int16)bitm(regv(rm), 16, 31) + (int16)bitm(regv(rn), 16, 31);
-	uint32 H = signed_sat(resH, 16, &issat);
+	uint16 H = signed_sat(resH, 16, &issat);
 
 	//结果综合
 	regv(rd) = (H << 16) | L;
